+++ clinicadl/callbacks/implemented/training_checkpoint.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 import logging
 from pathlib import Path
-from typing import TYPE_CHECKING, Any, Mapping
+from typing import TYPE_CHECKING
 
 import torch
 from pydantic import NonNegativeInt
