+++ clinicadl/train/__init__.py
@@ -1,8 +1,7 @@
 """
-To a model with ``ClinicaDL``'s trainer.
+To train/evaluate/predict with a model with ``ClinicaDL``'s trainer.
 """
 
 from .computational import ComputationalConfig
-
-# from .trainer import Trainer
+from .trainer import Trainer
 from .trainer_state import TrainerState
